refactor(banners): removing old banners for new composable ones
